Sharpfin Webserver 0.9 initial development

+++ src/ipk/sharpfin-www-ipk/files/mnt/debug/webserver/cgi-bin/admin_Tools_Backup.cgi
@@ -2,6 +2,7 @@
 
 DESTNAME="`/opt/webserver/cgi-bin/getarg destname`"
 PARTITION="`/opt/webserver/cgi-bin/getarg partition`"
+FAILURL="http://www.pschmidt.it/sharpfin/sharpfin-nanddump.install"
 
 #
 # Show Form
@@ -26,25 +27,35 @@ if [ ! "$DESTNAME" = "" ]; then
   echo "<br/>"
   echo "<p>Backup is now complete.  You can download this partition from "
   echo "<a href=\"/tmp/mtd-backup/$DESTNAME\">here</a></p>"
- 
 fi
 
 echo "<h1>Backup a Partition</h1>"
-echo "<p>The partitions are laid out on the NAND Flash as follows</p>"
-echo "<pre>"
-cat /proc/mtd
-echo "</pre>"
-echo "<p>Select one of the following links to create a backup file for one"
-echo "of the partitions.  This backup file can be restored using 'sharpflash',"
-echo "which is a JTAG programming utility.</p>"
+which nanddump >/dev/null 2>&1
+if [ "$?" -eq 0 -o -f "/usr/sbin/nanddump" ]
+then
+  echo "<p>The partitions are laid out on the NAND Flash as follows</p>"
+  echo "<pre>"
+  cat /proc/mtd
+  echo "</pre>"
+  echo "<p>Select one of the following links to create a backup file for one"
+  echo "of the partitions.  This backup file can be restored using 'sharpflash',"
+  echo "which is a JTAG programming utility.</p>"
+  
+  echo "<ul>"
+  echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=0&destname=boot-mtd.bin'>boot partition (mtd0)</a></li>"
+  echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=1&destname=kernel-mtd.bin'>kernel partition (mtd1)</a></li>"
+  echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=2&destname=root-mtd.bin'>root partition (mtd2)</a></li>"
+  echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=3&destname=config-mtd.bin'>config partition (mtd3)</a></li>"
+  echo "<li><a href=/cgi-bin/admin_Tools_Backup.cgi?partition=4&destname=debug-mtd.bin'>debug partition (mtd4)</a></li>"
+  if [ -c "/dev/mtd/5" ]
+  then
+    echo "<li><a href=/cgi-bin/admin_Tools_Backup.cgi?partition=5&destname=data-mtd.bin'>data partition (mtd5)</a></li>"
+  fi
 
-echo "<ul>"
-echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=0&destname=boot-mtd.bin'>boot partition (mtd0)</a></li>"
-echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=1&destname=kernel-mtd.bin'>kernel partition (mtd1)</a></li>"
-echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=2&destname=root-mtd.bin'>root partition (mtd2)</a></li>"
-echo "<li><a href='/cgi-bin/admin_Tools_Backup.cgi?partition=3&destname=config-mtd.bin'>config partition (mtd3)</a></li>"
-echo "<li><a href=/cgi-bin/admin_Tools_Backup.cgi?partition=4&destname=debug-mtd.bin'>debug partition (mtd4)</a></li>"
-echo "</ul>"
-
+  echo "</ul>"
+else
+  echo "<h1>WARNING:</h1>"
+  echo "To backup a partition the <i>nanddump</i> utility is needed, but it was NOT yet installed on your radio."
+  echo "To install the missing utility click <a href='/cgi-bin/admin_Install_Install_Addon.cgi?url="$FAILURL"&action=url'>here</a>"
+fi
 echo "</body></html>"
-
